Move command, validator package to controller/form, controller/validator
