+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/controller/PlayerSettingsController.java
@@ -27,7 +27,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import com.tesshu.jpsonic.command.PlayerSettingsCommand;
+import com.tesshu.jpsonic.controller.form.PlayerSettingsCommand;
 import com.tesshu.jpsonic.feature.auth.jwt.JWTAuthenticationToken;
 import com.tesshu.jpsonic.infrastructure.settings.SKeys;
 import com.tesshu.jpsonic.infrastructure.settings.SettingsFacade;
