+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/controller/UserSettingsController.java
@@ -28,7 +28,8 @@ import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-import com.tesshu.jpsonic.command.UserSettingsCommand;
+import com.tesshu.jpsonic.controller.form.UserSettingsCommand;
+import com.tesshu.jpsonic.controller.validator.UserSettingsValidator;
 import com.tesshu.jpsonic.domain.system.TranscodeScheme;
 import com.tesshu.jpsonic.infrastructure.settings.SKeys;
 import com.tesshu.jpsonic.infrastructure.settings.SettingsFacade;
@@ -43,7 +44,6 @@ import com.tesshu.jpsonic.service.ShareService;
 import com.tesshu.jpsonic.service.TranscodingService;
 import com.tesshu.jpsonic.util.PlayerUtils;
 import com.tesshu.jpsonic.util.StringUtil;
-import com.tesshu.jpsonic.validator.UserSettingsValidator;
 import jakarta.servlet.http.HttpServletRequest;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Controller;
