+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/controller/form/PersonalSettingsCommand.java
@@ -19,7 +19,7 @@
  * (C) 2018 tesshucom
  */
 
-package com.tesshu.jpsonic.command;
+package com.tesshu.jpsonic.controller.form;
 
 import java.util.List;
