+++ jpsonic-main/src/main/webapp/WEB-INF/jsp/databaseSettings.jsp
@@ -1,5 +1,5 @@
 <%@ page language="java" contentType="text/html; charset=utf-8" pageEncoding="iso-8859-1" %>
-<%--@elvariable id="command" type="com.tesshu.jpsonic.command.DatabaseSettingsCommand"--%>
+<%--@elvariable id="command" type="com.tesshu.jpsonic.controller.form.DatabaseSettingsCommand"--%>
 
 <html>
 <head>
