+++ jpsonic-main/src/main/webapp/WEB-INF/jsp/podcastSettings.jsp
@@ -1,5 +1,5 @@
 <%@ page language="java" contentType="text/html; charset=utf-8" pageEncoding="iso-8859-1" %>
-<%--@elvariable id="command" type="com.tesshu.jpsonic.command.PodcastSettingsCommand"--%>
+<%--@elvariable id="command" type="com.tesshu.jpsonic.controller.form.PodcastSettingsCommand"--%>
 
 <html><head>
     <%@ include file="head.jsp" %>
