+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/controller/PlayerSettingsControllerTest.java
@@ -27,7 +27,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.ExecutionException;
 
-import com.tesshu.jpsonic.command.PlayerSettingsCommand;
+import com.tesshu.jpsonic.controller.form.PlayerSettingsCommand;
 import com.tesshu.jpsonic.infrastructure.settings.SettingsFacadeBuilder;
 import com.tesshu.jpsonic.persistence.api.entity.Player;
 import com.tesshu.jpsonic.persistence.api.entity.Transcoding;
