+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/controller/PodcastSettingsControllerTest.java
@@ -23,7 +23,7 @@ import static com.tesshu.jpsonic.service.ServiceMockUtils.mock;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import com.tesshu.jpsonic.command.PodcastSettingsCommand;
+import com.tesshu.jpsonic.controller.form.PodcastSettingsCommand;
 import com.tesshu.jpsonic.infrastructure.settings.SKeys;
 import com.tesshu.jpsonic.infrastructure.settings.SettingsFacade;
 import com.tesshu.jpsonic.infrastructure.settings.SettingsFacadeBuilder;
